refactor: 로또 메시지 LottoMessage 클래스로 분리
