+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -25,14 +25,6 @@ class LottoTest {
 	}
 
 	@Test
-	void 로또_번호_문자열_반환() {
-		assertThat(Lotto.inputNumber(Arrays.asList(1, 2, 3, 4, 5, 6)).getResultMessage()).isEqualTo(
-			"[1, 2, 3, 4, 5, 6]");
-		assertThat(Lotto.inputNumber(Arrays.asList(6, 5, 4, 3, 2, 1)).getResultMessage()).isEqualTo(
-			"[1, 2, 3, 4, 5, 6]");
-	}
-
-	@Test
 	void 로또_번호_포함_여부() {
 		assertThat(Lotto.inputNumber(Arrays.asList(1, 2, 3, 4, 5, 6)).contains(LottoNumber.from(1))).isTrue();
 		assertThat(Lotto.inputNumber(Arrays.asList(1, 2, 3, 4, 5, 6)).contains(LottoNumber.from(7))).isFalse();
